Add leaderboard route
I’m not sure exactly what the issue was that you were running into but
it seems like you have all the code to make this work already written.

I uncommented lines 4-5, and moved line 5 to the end of the end of the
method. The call to render the the leaderboard template (`erb
:leaderboard`) needs to come last so that it is the return value of
this method.

Now the page loads!