User an Array for leaderboard`
I’d recommend using an Array for the leaderboard so that it will be
easier to sort and work with within the view (HTML).

Using `#uniq!` to make sure that we don’t have duplicates is probably
not the best possible way to do this but we’ll get to that a little
later.

We can use `#uniq!` instead of `#uniq` to permanently change the value
of the array rather than just returning the value of the unique array.
This way we don’t need to create another variable to store the unique
array.

After these changes, leaderboard will produce a data structure that
looks like this:

```ruby
[
  {:name=>"Patriots", :win=>0, :loss=>0},
  {:name=>"Broncos", :win=>0, :loss=>0},
  {:name=>"Colts", :win=>0, :loss=>0},
  {:name=>"Steelers", :win=>0, :loss=>0}
]
```

Also, this change will break the view, but we'll fix that in the next
step.