Use #find to find teams
Since our teams are now stored in an Array instead of a hash, we’ve
made it a little bit more complicated to find a given team out of the
`leaderboard`. We’ll refactor this a little but in the next step but
basically this will iterate through the `leaderboard ` Array until it
finds a team that has a name that matches either the
`match[:away_team]` or `match[:home_team]`, depending on which we’re
looking for.

I also changed the sorting a little bit so that you don’t need to call
reverse on it.