Add #find_team
We can move the logic for finding a team into a method. This way, if we
ever change the way that we need to find a team, we will only need to
update our logic in one place. It also reads nicer now that the method
name kind of says exactly what the code does. Since the method name
describes the code well, we don’t need to read into how it works unless
we want to. We can read lines 50-51, 56-57 without having to try too
hard to understand what’s going on.